Move ActivationTxHeader to its own file to avoid scale code gen for it
